Ignore from/until URL parameters in v3 of protocol
In version 3 of the protocol, it's valid to send a request entirely in
the HTTP body and have no URL parameters. The previous versions of these
patches didn't handle this case correctly, and errored out once they
failed to parse integers from the empty string strings that looking up
"from" and "until" gave.

In this patch, we choose to deal with this situation by saying that for
a request made with version 3 of the protocol, we _only_ look at the
data that was passed in the HTTP body when looking for targets. For
other versions of the protocol, we _only_ look at the data passed in the
URL parameters. This is how things used to work before as well, c.f.:

    git show 9a28449:carbonserver/render.go

We also split the target parsing into its own function and unify some
error handling.