Update go-graphite/protocol dependency
We need the new version of v3 to fix an issue with timeShifted metrics.

Really weirdly, go-graphite/protocol was not in Gopkg.toml but only in
Gopkg.lock. This suggests it was pulled in as a transitive dependency,
in which case it could only have come from go-graphite/carbonzipper, but
that is also only in Gopkg.lock. Maybe this is a problem with dep and
how it deals with packages where we only really depend on a subpackage?