Write target names to log
Zap doesn't like when we write a map to our logs. Instead of doing
something more complicated, we just write the names of the metrics we
were asked to fetch.

This more or less reverts the logging behaviour to before commit
72241bd. The difference is that we used to log from/until times as well.
That's trickier to do accurately now, as a requests with version 3 of
the protocol may include more than one from/until pair.