Hard max creates per second
A new option

    hard-max-creates-per-second = false

that's off by default makes the max-creates-per-second limit added in
d315e96 into a hard limit, meaning that we drop any new metrics above
the limit instead of queueing them for later.

This is desired behaviour when operating in an environment where we have
multiple independent concurrent producers of new metrics, and some of
them may flood us with new metrics by accident. Having a soft limit in
that case may cause the store to run out of memory and be killed. It is
preferable to impose a hard rate limit on everyone in that case, until
the offender can be found and convinced to cease their flooding. Once
that is done, new metrics that were previously dropped will slowly be
accepted as new data points are written to them.

There's an unintended ramp-up period to when the limiting starts taking
effect, as we only begin load-shedding metrics at the prescribed rate
after we've accepted more than our share for a little while. I think
this is because time.Ticker claims to try to adjust for slow receivers.
It doesn't seem very important, as the desired limiting and
load-shedding kicks in after a few seconds.