Stop leaking goroutines in hard throttler
After rolling out 8016b45 we eventually realized that we were leaking
goroutines at the rate of 1 per second. I'm not sure what the problem
with the old implementation was, as I thought the cancel() call at the
beginning of every loop iteration would prevent these leaks, but we can
get rid of the problem by not spawning any goroutines at all by
following the implementation of the soft throttler more closely.