Correctly disable soft rate-limiting
When we disable soft rate-limiting, we close the channel our ticker
sends messages on. If we try to receive from a closed channel, we get
the default zero value for the type we expected. The logic we added in
8016b45 interprets this as saying we are hard-limiting the creation of
metrics, and drops any new metrics.

The fix we propose here is to check whether the channel we're receiving
from is closed. If it is, we skip the rate-limiting checks.

This fixes issue https://github.com/lomik/go-carbon/issues/247.