Label Prometheus metrics by config labels
The Prometheus Go client documentation [1] says:

    [The function we're using to add fixed labels to all metrics
    exposed] should not be used to add fixed labels to all metrics
    exposed.

The Prometheus documentation [2] also says:

    If you ever find yourself wanting to apply the same label to all of
    your metrics, stop.

    [...]

    The second case is when a label is really a target label. These are
    things like region, cluster names, and so on, that come from your
    infrastructure setup rather than the application itself. It’s not
    for an application to say where it fits in your label taxonomy,
    that’s for the person running the Prometheus server to configure and
    different people monitoring the same application may give it
    different names.

    Accordingly, these labels belong up in the scrape configs of
    Prometheus via whatever service discovery you’re using. It’s okay to
    apply the concept of machine roles here as well, as it’s likely
    useful information for at least some people scraping it.

In this commit, we're ignoring both of these suggestions and adding
labels to all the metrics we expose in the instrumentation code itself.

We're doing that because we want to label all the metrics coming from a
given go-carbon instance with information about the network topology it
lives in; things like its cluster and ring position. That information
would be tricky to feed to the Prometheus scraping configuration, while
being something the go-carbon instance itself can easily know. We thus
feel that this is enough of a corner case to ignore the official advice.

[1] https://godoc.org/github.com/prometheus/client_golang/prometheus#WrapRegistererWith
[2] https://prometheus.io/docs/instrumenting/writing_exporters/#target-labels-not-static-scraped-labels