Fix pubsub test. Simplify test target in Makefile.
While working on switching to modules I noticed that we don't need to
specify each package as target in Makefile 'test'.

When trying to replace that target with ./..., it started hanging
because of pubsub tests. By fixing this test, we can simplify Makefile.

Pubsub tests had  two issues:
 * hanging because we were checking if a context was canceled on
the error returned from rcv.subscription.Receive instead of calling
Err() on the context.
 * invalid test data(probably parsing got improved over time and the
   test was not executed)