systemd: After and Wants network-online and local-fs
Without these go-carbon can start before the network is online
and also before all local fses have been mounted. If /var/lib/carbon
is on a separate disk, this can lead to a failed start because
the go-carbon will start, begin writing to the empty dir, then
the carbon FS will not mount due to it not being empty.

Signed-off-by: Dan van der Ster <daniel.vanderster@cern.ch>