receiver/kafka: fix kafka connection leaks when failing to consume partition
Clients and consumers are not closed when consumer.ConsumePartition failed, which
creates a resource leaks. If the error is sticky due to bad offset saved in state
file, the receiver would keep creating new consumer connections and not closing them.

This commit fixes the issue, by making sure the connections and consumers are
closed if there is errors when trying to consume a partition.

At the same time, if an sarama.ErrOffsetOutOfRange error is returned, it would just
falls back to oldest offset, instead of keep retrying.