carbonserver: change max-metrics-globbed and max-metrics-rendered defaults
The default values might be too low for some of the users and it is a default introduced mainly
for trie index. This commit change the default to a much bigger value to avoid silently truncated
query result.

At the same time, logging truncated errors to access.log to improve visibility.