quota: adopt a more performant throughput throttling implementation
When it comes to point search, map lookup is very fast, much faster than traversing the trie tree.
What's more, the current trie implementation adopts a space efficient strategy and children
are not sorted. And this means much higher cpu usage. We are seeing up to 700% cpu usage growth
in a cluster receving 1 millions+ data points.

By adopting a map-based implementation, we are able to cut down up to 600% of cpu usage in
that cluster, which means it is much more efficient to collect and control throughput usage.