trie: stop indexing empty directory nodes
There is a strange bug in trie that seems related to indexing empty directory nodes. The actual root cause
is still unknown to me and I fail to reproduce the issue. In our production, we notice the issue associates
with a cron job for removing files (which would causes the generation of empty directories).

Considering that empty directories doesnt produce much value in go-carbon/graphite, excluding it from
index should cause no harm. And this change also appears to resolve the issue on our production.