quota: realtime usage update for metrics, logical/physical sizes, and data points
Without this patch, it is possible for some namespaces to have a window of 1-3 `carbonserver.quota-usage-report-frequency`
to conduct quota enforcement penetration.

There are two cases that need to be addressed, dir nodes that are already annotated with proper quota config, nodes that
don't have quota config due to realtime insert via *CarbonserverListener.newMetricsChan.

For annotated nodes, trie index resovle it via realtime metric counter update in *trieIndex.insert.

For unannotated nodes, a new config are introduced, which is called transient-child-limit. With this config, during realtime
insert, trieIndex would annotate the new dir nodes of parent nodes that enables that config. This is not the most proper way
to address the issue, but a trade off of correctness, reliability, complexity, and performance. Re-iterating all quota configs
for every realtime insert is nice but expensive. However, as always, I could be wrong. But this seems to be working in my tests.