Use find cache for glob expansion in grpc render
The reason is that caching render responses in streaming render
doesn't make much sense. Also, it reduces performance cause the
results get cached when the streaming finishes. All this, seems
unnecessary if we could cache glob expansion because most of the
times that is what makes requests long (from the trace observation).